divide vote in votecontroller
